backend/requirements.txt
flask
python-dotenv

#Para dps

#Flask-JWT-Extended
#Flask-SQLAlchemy
#marshmallow
